pyproject.toml
[tool.poetry]
name = "pdf-interpreter"
version = "0.1.0"
description = ""
authors = ["takumi-nishimura <clp13218@nitech.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
langchain-community = "^0.0.33"
langchain-openai = "^0.1.3"
qdrant-client = "^1.8.2"
pdfminer-six = "^20231228"
streamlit = "^1.33.0"
streamlit-javascript = "^0.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
